+++ ISD93xx_NAU85L40_1C_MIC_48kHz_DataOut/ISD93xx_NAU85L40_1C_MIC/ISD93xx_NAU85L40_1C/isd9340_keil/sample/driver/smpl_drvi2s_nau85L40_1C/delay_sound.h
@@ -3,30 +3,30 @@
 
 typedef struct delay {
 	int mic;
-	double *read_pointer;
-	double *write_pointer;
-	double num_samples_to_delay;
-	double fractional_delay;
-	double delay_blend_param;
-	double delay_tap_mix_volume;
-	double feedback_volume;
-	double thiran_coeff_1;
-	double thiran_coeff_2;
-	double thiran_coeff_3;
-	double prev_in[3];
-	double prev_out[3];
-	double *buffer;
+	float *read_pointer;
+	float *write_pointer;
+	float num_samples_to_delay;
+	float fractional_delay;
+	float delay_blend_param;
+	float delay_tap_mix_volume;
+	float feedback_volume;
+	float thiran_coeff_1;
+	float thiran_coeff_2;
+	float thiran_coeff_3;
+	float prev_in[3];
+	float prev_out[3];
+	float *buffer;
 } delay;
 
 
-extern delay * delay_init(double num_samples_to_delay, double fractional_delay, double fb_v, double dtm_v,double blend_param, int mic);
-extern void delay_set_delay(delay *d, double num_samples_to_delay, double fractional_delay);
-void delay_set_feedback_volume(delay *d, double fb_v);
-void delay_set_delay_tap_mix_volume(delay *d, double dtm_v);
+extern delay * delay_init(float num_samples_to_delay, float fractional_delay, float fb_v, float dtm_v,float blend_param, int mic);
+extern void delay_set_delay(delay *d, float num_samples_to_delay, float fractional_delay);
+void delay_set_feedback_volume(delay *d, float fb_v);
+void delay_set_delay_tap_mix_volume(delay *d, float dtm_v);
 void delay_set_mic(delay *d, int mic);
-void delay_set_blend_param(delay *d, double delay_blend_param);
-extern double delay_out(delay *d, double in);
-//extern void change_delay(double angle);
+void delay_set_blend_param(delay *d, float delay_blend_param);
+extern float delay_out(delay *d, float in);
+//extern void change_delay(float angle);
 
 
 #endif
