+++ ISD93xx_NAU85L40_1C_MIC_48kHz_DataOut/ISD93xx_NAU85L40_1C_MIC/ISD93xx_NAU85L40_1C/isd9340_keil/sample/driver/smpl_drvi2s_nau85L40_1C/localization.h
@@ -3,20 +3,20 @@
 #define __LOCALIZATION_H__
 
 typedef struct microphone {
-	double x;
-	double y;
-	double * buffer;
+	float x;
+	float y;
+	float * buffer;
 } microphone;
 
-extern microphone * build_mic(double x, double y);
-extern double find_source(microphone * m1, microphone * m2, microphone * m3);
+extern microphone * build_mic(float x, float y);
+extern float find_source(microphone * m1, microphone * m2, microphone * m3);
 extern void set_buffer_size(int bf_size);
-double distance_calc(microphone * m1, microphone * m2);
-double degree_calc(microphone * m1, microphone * m2);
+float distance_calc(microphone * m1, microphone * m2);
+float degree_calc(microphone * m1, microphone * m2);
 extern void init_triangle(microphone * m1, microphone * m2, microphone * m3); 
-extern void set_buffer(microphone * m1, double * buffer);
+extern void set_buffer(microphone * m1, float * buffer);
 void find_mic_3(microphone * m1, microphone * m2);
-double calc_right(double delay, double distance, double angle);
-double calc_left(double delay, double distance, double angle);
-double to_degrees(double angle);
+float calc_right(float delay, float distance, float angle);
+float calc_left(float delay, float distance, float angle);
+float to_degrees(float angle);
 #endif
